% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.oolax}
\alias{anova.oolax}
\title{Comparison of nested models}
\usage{
\method{anova}{oolax}(object, object2, ...)
}
\arguments{
\item{object}{An object of class \code{"oolax"}, returned by
\code{\link{alogLik}} or \code{\link{adjust_object}}.}

\item{object2}{An object of class \code{"chandwich"}, returned by
\code{\link{alogLik}} or \code{\link{adjust_object}}.}

\item{...}{Further objects of class \code{"oolax"} and/or arguments
to be passed to \code{\link[chandwich]{anova.chandwich}}.}
}
\value{
An object of class \code{"anova"} inheriting from class
 \code{"data.frame"}, with four columns:
    \item{Model.Df}{The number of parameters in the model}
    \item{Df}{The decrease in the number of parameter compared the model
      in the previous row}
    \item{ALRTS}{The adjusted likelihood ratio test statistic}
    \item{Pr(>ALRTS)}{The p-value associated with the test that the
      model is a valid simplification of the model in the previous row.}
 The row names are the names of the model objects.
}
\description{
\code{anova} method for objects of class \code{"oolax"}.
Compares two or more nested models using the adjusted likelihood ratio
test statistic (ALRTS) described in Section 3.5 of
\href{http://dx.doi.org/10.1093/biomet/asm015}{Chandler and Bate (2007)}.
The nesting must result from the simple constraint that a subset of the
parameters of the larger model is held fixed.
}
\details{
The objects of class \code{"oolax"} need not be provided in nested
  order: they will be ordered inside \code{anova.oolax} based on the
  values of \code{attr(., "p_current")}.
}
\examples{
got_evd <- requireNamespace("evd", quietly = TRUE)

if (got_evd) {
  y <- c(chandwich::owtemps[, "Oxford"], chandwich::owtemps[, "Worthing"])
  x <- rep(c(1, -1), each = length(y) / 2)
  owfit <- evd::fgev(y, nsloc = x)
  year <- rep(1:(length(y) / 2), 2)
  oola_small <- alogLik(owfit, cluster = year)

  owfit <- evd::fgev(y)
  year <- rep(1:(length(y) / 2), 2)
  oola_tiny <- alogLik(owfit, cluster = year)

  anova(oola_small, oola_tiny)
}
}
\references{
Chandler, R. E. and Bate, S. (2007). Inference for clustered
  data using the independence loglikelihood. \emph{Biometrika},
  \strong{94}(1), 167-183. \url{http://dx.doi.org/10.1093/biomet/asm015}
}
\seealso{
\code{\link[chandwich]{anova.chandwich}}.

\code{\link{adj_object}}.
}
