% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bernoulli}
\alias{bernoulli}
\alias{bernoulli_loglik}
\alias{fit_bernoulli}
\alias{nobs.bernoulli}
\alias{coef.bernoulli}
\alias{vcov.bernoulli}
\alias{logLik.bernoulli}
\title{Inference for the Bernoulli distribution}
\usage{
bernoulli_loglik(prob, data)

fit_bernoulli(data)

\method{nobs}{bernoulli}(object, ...)

\method{coef}{bernoulli}(object, ...)

\method{vcov}{bernoulli}(object, ...)

\method{logLik}{bernoulli}(object, ...)
}
\arguments{
\item{prob}{A numeric scalar. Probability of success on each trial.}

\item{data}{A numeric vector of outcomes from Bernoulli trials: 0 for a
failure, 1 for a success.  Alternatively, a logical vector with FALSE
for a failure and TRUE for a success.}
}
\value{
\code{fit_bernoulli} returns an object of class \code{"bernoulli"}, a list
with components: \code{data, logLik, mle, nobs, vcov}, where \code{data}
are the input data.

\code{bernoulli_loglik} returns an object of class \code{"logLikVec"}, a
vector length \code{length(data)} containing the likelihood contributions
from the individual observations in \code{data}.
}
\description{
Functions involved in making inferences about the probability of success
in a Bernoulli distribution.
}
\details{
\code{fit_bernoulli}: fit a Bernoulli distribution

\code{bernoulli_loglik}: calculates contributions to a log-likelihood based
on the Bernoulli distribution.  The log-likelihood is calculated up to an
additive constant.

\code{nobs, coef, vcov} and \code{logLik} methods are provided.
}
\examples{
# Set up data
x <- exdex::newlyn
u <- quantile(x, probs = 0.6)
exc <- x > u

# Fit a Bernoulli distribution
fit <- fit_bernoulli(exc)

# Calculate the log-likelihood at the MLE
res <- bernoulli_loglik(fit$mle, exc)

# The logLik method sums the individual log-likelihood contributions.
logLik(res)

# nobs, coef, vcov, logLik methods for objects returned from fit_bernoulli()
nobs(fit)
coef(fit)
vcov(fit)
logLik(fit)
}
\seealso{
\code{\link[stats]{Binomial}}.  The Bernoulli distribution is the
  special case where \code{size = 1}.
}
