% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evd.R
\name{evd}
\alias{evd}
\alias{alogLik.evd}
\title{Loglikelihood adjustment of evd fits}
\usage{
\method{alogLik}{evd}(x, cluster = NULL, use_vcov = TRUE, ...)
}
\description{
S3 \code{alogLik} method to perform loglikelihood adjustment of fitted
extreme value model objects produced by the evd package.
}
\details{
These functions are not intended to be called by the user.
}
\examples{
# We need the evd package
got_evd <- requireNamespace("evd", quietly = TRUE)

if (got_evd) {
  library(evd)
  # An example from the evd::fgev documentation
  uvdata <- evd::rgev(100, loc = 0.13, scale = 1.1, shape = 0.2)
  M1 <- evd::fgev(uvdata, nsloc = (-49:50)/100)
  adj_fgev <- alogLik(M1)
  summary(adj_fgev)

  # An example from Chandler and Bate (2007)
  y <- c(chandwich::owtemps[, "Oxford"], chandwich::owtemps[, "Worthing"])
  x <- rep(c(1, -1), each = length(y) / 2)
  owfit <- evd::fgev(y, nsloc = x)
  year <- rep(rownames(chandwich::owtemps), 2)
  adj_owfit <- alogLik(owfit, cluster = year)
  summary(adj_owfit)

  # An example from the evd::fpot documentation
  uvdata <- evd::rgpd(100, loc = 0, scale = 1.1, shape = 0.2)
  M1 <- evd::fpot(uvdata, 1)
  adj_fpot <- alogLik(M1)
  summary(adj_fpot)
  # Fit using the pp model, rather than the gpd
  M1 <- evd::fpot(uvdata, 1, model = "pp", npp = 365)
  adj_fpot <- alogLik(M1)
  summary(adj_fpot)
}
}
