% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evir.R
\name{evir}
\alias{evir}
\alias{alogLik.gev}
\alias{alogLik.gpd}
\alias{alogLik.potd}
\title{Loglikelihood adjustment of evir fits}
\usage{
\method{alogLik}{gev}(x, cluster = NULL, use_vcov = TRUE, ...)

\method{alogLik}{gpd}(x, cluster = NULL, use_vcov = TRUE, ...)

\method{alogLik}{potd}(x, cluster = NULL, use_vcov = TRUE, ...)
}
\description{
S3 \code{alogLik} method to perform loglikelihood adjustment of fitted
extreme value model objects produced by the evir package.
}
\details{
These functions are not intended to be called by the user.
}
\examples{
# We need the evir package
got_evir <- requireNamespace("evir", quietly = TRUE)
if (got_evir) {
  library(evir)
  # An example from the evir::gev documentation
  data(bmw)
  out <- evir::gev(bmw, "month")
  adj_out <- alogLik(out)
  summary(adj_out)

  # An example from the evir::gpd documentation
  data(danish)
  out <- evir::gpd(danish, 10)
  adj_out <- alogLik(out)
  summary(adj_out)

  # An example from the evir::pot documentation
  # We use lax::oopot() to return the input data
  out <- oopot(danish, 10)
  adj_out <- alogLik(out)
  summary(adj_out)
}
}
