% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fExtremes.R
\name{fExtremes}
\alias{fExtremes}
\alias{alogLik.fGEVFIT}
\alias{alogLik.fGPDFIT}
\title{Loglikelihood adjustment of fExtremes fits}
\usage{
\method{alogLik}{fGEVFIT}(x, cluster = NULL, use_vcov = TRUE, ...)

\method{alogLik}{fGPDFIT}(x, cluster = NULL, use_vcov = TRUE, ...)
}
\description{
S3 \code{alogLik} method to perform loglikelihood adjustment of fitted
extreme value model objects produced by the
\code{\link[fExtremes:00Extremes-package]{fExtremes}} package.
}
\details{
These functions are not intended to be called by the user.
}
\examples{
# We need the fExtremes package
got_fExtremes <- requireNamespace("fExtremes", quietly = TRUE)
if (got_fExtremes) {
  library(fExtremes)

  # GEV
  # An example from the fExtremes::gevFit documentation
  x <- fExtremes::gevSim(model = list(xi=0.25, mu=0, beta=1), n = 1000)
  # Fit GEV distribution by maximum likelihood estimation
  fit <- gevFit(x)
  adj_fit <- alogLik(fit)
  summary(adj_fit)

  # GP
  # An example from the fExtremes::gpdFit documentation
  # Simulate GP data
  x <- gpdSim(model = list(xi = 0.25, mu = 0, beta = 1), n = 1000)
  # Fit GP distribution by maximum likelihood estimation
  fit <- gpdFit(x, u = min(x))
  adj_fit <- alogLik(fit)
  summary(adj_fit)
}
}
