% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismev_gpd_refit.R
\name{gpd_refit}
\alias{gpd_refit}
\title{Maximum-likelihood Fitting for the GPD Distribution}
\usage{
gpd_refit(xdat, threshold, npy = 365, ydat = NULL, sigl = NULL,
  shl = NULL, siglink = identity, shlink = identity,
  siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", maxit = 10000, ...)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted.}

\item{threshold}{The threshold; a single number or a numeric
    vector of the same length as \code{xdat}.}

\item{npy}{The number of observations per year/block.}

\item{ydat}{A matrix of covariates for generalized linear modelling
    of the parameters (or \code{NULL} (the default) for stationary
    fitting). The number of rows should be the same as the length
    of \code{xdat}.}

\item{sigl}{Numeric vectors of integers, giving the columns
    of \code{ydat} that contain covariates for generalized linear
    modelling of the scale and shape parameters repectively
    (or \code{NULL} (the default) if the corresponding parameter is
    stationary).}

\item{shl}{Numeric vectors of integers, giving the columns
    of \code{ydat} that contain covariates for generalized linear
    modelling of the scale and shape parameters repectively
    (or \code{NULL} (the default) if the corresponding parameter is
    stationary).}

\item{siglink}{Inverse link functions for generalized
    linear modelling of the scale and shape parameters repectively.}

\item{shlink}{Inverse link functions for generalized
    linear modelling of the scale and shape parameters repectively.}

\item{siginit}{numeric giving initial value(s) for parameter
	estimates.  If NULL, default is sqrt(6 * var(xdat))/pi and 0.1
	for the scale and shape parameters, resp.  If using parameter
	covariates, then these values are used for the constant term,
	and zeros for all other terms.}

\item{shinit}{numeric giving initial value(s) for parameter
	estimates.  If NULL, default is sqrt(6 * var(xdat))/pi and 0.1
	for the scale and shape parameters, resp.  If using parameter
	covariates, then these values are used for the constant term,
	and zeros for all other terms.}

\item{show}{Logical; if \code{TRUE} (the default), print details of
    the fit.}

\item{method}{The optimization method (see \code{\link{optim}} for
    details).}

\item{maxit}{The maximum number of iterations.}

\item{...}{Other control parameters for the optimization. These
    are passed to components of the \code{control} argument of
    \code{optim}.}
}
\description{
This is a slightly modified version of the \code{\link[ismev]{gpd.fit}}
function in the \code{\link[ismev]{ismev}} package.
The modification is to add to the returned object the arguments
\code{xdat, ydat, siglink, shlink} and matrices
\code{sigmat, shmat} giving the respective regression design matrices
for the scale and shape parameters of the model.
}
\examples{
# We need the ismev package
got_ismev <- requireNamespace("ismev", quietly = TRUE)
if (got_ismev) {
  library(ismev)
  data(rain)
  fit1 <- gpd.fit(rain, 10)
  ls(fit1)
  fit2 <- gpd_refit(rain, 10)
  ls(fit2)
}
}
\references{
Heffernan, J. E. and Stephenson, A. G. (2018). ismev: An
  Introduction to Statistical Modeling of Extreme Values.
  R package version 1.42.
  \url{https://CRAN.R-project.org/package=ismev}.
}
