% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismev_oo_gpd_fit.R
\name{oogpd.fit}
\alias{oogpd.fit}
\title{Maximum-likelihood Fitting for the GPD Distribution}
\usage{
oogpd.fit(xdat, threshold, npy = 365, ydat = NULL, sigl = NULL,
  shl = NULL, siglink = identity, shlink = identity,
  siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", maxit = 10000, ...)
}
\description{
This is a slightly modified version of the \code{\link[ismev]{gpd.fit}}
function in the \code{\link[ismev]{ismev}} package.
The modification is to add to the returned object the arguments
\code{xdat, ydat, siglink, shlink} and matrices
\code{sigmat, shmat} giving the respective regression design matrices
for the scale and shape parameters of the model.
}
\examples{
# We need the evd package
got_ismev <- requireNamespace("ismev", quietly = TRUE)
if (got_ismev) {
  data(rain)
  fit1 <- gpd.fit(rain, 10)
  ls(fit1)
  fit2 <- oogpd.fit(rain, 10)
  ls(fit2)
}
}
\references{
Heffernan, J. E. and Stephenson, A. G. (2018). ismev: An
  Introduction to Statistical Modeling of Extreme Values.
  R package version 1.42.
  \url{https://CRAN.R-project.org/package=ismev}.
}
