% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolax.R
\docType{package}
\name{oolax}
\alias{oolax}
\alias{oolax-package}
\title{oolax: Objected-orientated Loglikelihood Adjustment for Extreme Value Models}
\description{
Performs adjusted inferences using fitted model objects returned from the
extreme value analysis packages
\code{evd} \url{https://cran.r-project.org/package=evd},
\code{ismev} \url{https://cran.r-project.org/package=ismev}, ...
Adjusted standard errors and an adjusted loglikelihood are provided.
The adjustment is based on a robust sandwich estimator of the parameter
covariance matrix, based on the methodology in Chandler and Bate (2007)
<doi:10.1093/biomet/asm015>. This can be used for cluster correlated data
when interest lies in the parameters of the marginal distributions or for
performing inferences that are robust to certain types of model
misspecification.  Currently only univariate extreme value models are
supported.  As a byproduct, generic S3 methods logLik, coef, vcov and nobs
are provided, where these are not already provided in the original package.
}
\details{
\code{\link[oolax]{evd}},
\code{\link[oolax]{ismev}}.

See Chandler and Bate (2007) for full details and
\code{vignette("oola-vignette", package = "oola")} for an
overview of the package.
}
\references{
Chandler, R. E. and Bate, S. (2007). Inference for clustered
  data using the independence loglikelihood. \emph{Biometrika},
  \strong{94}(1), 167-183. \url{http://dx.doi.org/10.1093/biomet/asm015}
}
