% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evir_oo_pot.R
\name{oopot}
\alias{oopot}
\title{Fits a Poisson point process to the data, an approach sometimes known as
peaks over thresholds (POT), and returns an object of class "potd".}
\usage{
oopot(data, threshold = NA, nextremes = NA, run = NA,
  picture = TRUE, ...)
}
\description{
This is a slightly modified versions of \code{\link[evir]{pot}}.
The main modification is to add to the returned object the argument
\code{data} supplied by the user.  This is added to the returned
(list) object with the name \code{input_data}.
}
\examples{
# We need the evd package
got_evir <- requireNamespace("evir", quietly = TRUE)
if (got_evir) {
  data(danish)
  out <- oopot(danish, 10)
  ls(out)
}
}
