% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismev_oo_pp_fit.R
\name{oopp.fit}
\alias{oopp.fit}
\title{Maximum-likelihood Fitting for the Point Process Model}
\usage{
oopp.fit(xdat, threshold, npy = 365, ydat = NULL, mul = NULL,
  sigl = NULL, shl = NULL, mulink = identity, siglink = identity,
  shlink = identity, muinit = NULL, siginit = NULL, shinit = NULL,
  show = TRUE, method = "Nelder-Mead", maxit = 10000, ...)
}
\description{
This is a slightly modified versions of the \code{\link[ismev]{gev.fit}}
function in the \code{\link[ismev]{ismev}} package.
The main modification is to add to the returned object the arguments
\code{xdat, ydat, mulink, siglink, shlink} and matrices
\code{mumat, sigmat, shmat} giving the respective regression design matrices
for the location, scale and shape parameters of the model.  In addition, a
bug in the code that sets initial estimates has been corrected: the bug
meant that if \code{threshold} is a vector then the optimization hangs.
}
\examples{
# We need the evd package
got_ismev <- requireNamespace("ismev", quietly = TRUE)
if (got_ismev) {
  data(rain)
  fit1 <- pp.fit(rain, 100, show = FALSE)
  ls(fit1)
  fit2 <- oopp.fit(rain, 10, show = FALSE)
  ls(fit2)
}
}
\references{
Heffernan, J. E. and Stephenson, A. G. (2018). ismev: An
  Introduction to Statistical Modeling of Extreme Values.
  R package version 1.42.
  \url{https://CRAN.R-project.org/package=ismev}.
}
