% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POT.R
\name{POT}
\alias{POT}
\alias{alogLik.uvpot}
\title{Loglikelihood adjustment of POT fits}
\usage{
\method{alogLik}{uvpot}(x, cluster = NULL, use_vcov = TRUE, ...)
}
\description{
S3 \code{alogLik} method to perform loglikelihood adjustment of fitted
extreme value model objects produced by the POT package.
}
\details{
These functions are not intended to be called by the user.
}
\examples{
# We need the evd package
got_POT <- requireNamespace("POT", quietly = TRUE)

if (got_POT) {
  library(POT)
  # An example from the POT::fitgpd documentation.
  x <- POT::rgpd(200, 1, 2, 0.25)
  fit <- POT::fitgpd(x, 1, "mle")
  adj_fit <- alogLik(fit)
}
}
