% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lax-internal.R
\name{lax-internal}
\alias{lax-internal}
\alias{adj_object}
\alias{ismev_ppp}
\alias{return_level_gev}
\alias{gev_rl_prof}
\alias{gev_rl_CI}
\alias{box_cox_deriv}
\title{Internal lax functions}
\usage{
adj_object(x, cluster = NULL, use_vcov = TRUE, ...)

ismev_ppp(a, npy)

return_level_gev(x, m, level, npy, prof, inc, type)

gev_rl_prof(x, m, level, npy, inc, type, rl_sym)

gev_rl_CI(x, m, level, npy, type)

box_cox_deriv(x, lambda = 1, lambda_tol = 1/50, poly_order = 3)
}
\description{
Internal lax functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
