% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alogLik_methods.R
\name{logLikVec}
\alias{logLikVec}
\title{Evaluate loglikelihood contributions from specific observations}
\usage{
logLikVec(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments.}
}
\value{
An object of class \code{"logLikVec"}, a vector containing
  individual loglikelihood contributions.
}
\description{
Generic function for calculating loglikelihood contributions from
individual observations for a fitted model.
}
\details{
See \code{\link{alogLik}}: loglikelihood adjustment for model fits.
}
\section{Examples}{

See the example in \code{\link{bernoulli}}.
}

