% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLikVec}
\alias{logLikVec}
\title{Evaluate loglikelihood contributions from specific observations}
\usage{
logLikVec(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments.}
}
\description{
Generic function for calculating loglikelihood contributions from
individual observations for a fitted model.
}
