% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo_gev_fit.R
\name{oogev.fit}
\alias{oogev.fit}
\title{Maximum-likelihood Fitting of the GEV Distribution}
\usage{
oogev.fit(xdat, ydat = NULL, mul = NULL, sigl = NULL, shl = NULL,
  mulink = identity, siglink = identity, shlink = identity,
  muinit = NULL, siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", maxit = 10000, ...)
}
\description{
This is a slightly modified version of \code{\link[ismev]{gev.fit}}.
The modification is to add to the returned object the arguments
\code{xdat, ydat, mulink, siglink, shlink} and matrices
\code{mumat, sigmat, shmat} giving the respective regression design matrices
for the location, scale and shape parameters of the model.
}
\examples{
if (got_ismev) {
  fit1 <- gev.fit(revdbayes::portpirie, show = FALSE)
  ls(fit1)
  fit2 <- oogev.fit(revdbayes::portpirie, show = FALSE)
  ls(fit2)
}
}
