% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo_pp_fit.R
\name{oopp.fit}
\alias{oopp.fit}
\title{Maximum-likelihood Fitting for the Point Process Model}
\usage{
oopp.fit(xdat, threshold, npy = 365, ydat = NULL, mul = NULL,
  sigl = NULL, shl = NULL, mulink = identity, siglink = identity,
  shlink = identity, muinit = NULL, siginit = NULL, shinit = NULL,
  show = TRUE, method = "Nelder-Mead", maxit = 10000, ...)
}
\description{
This is a slightly modified versions of \code{\link[ismev]{gev.fit}}.
The main modification is to add to the returned object the arguments
\code{xdat, mulink, siglink, shlink} and matrices
\code{mumat, sigmat, shmat} giving the respective regression design matrices
for the location, scale and shape parameters of the model.  In addition, a
bug in the code that sets initial estimates has been corrected: the bug
meant that if \code{threshold} is a vector then the optimization hangs.
}
\examples{
if (got_ismev) {
  data(rain)
  fit1 <- pp.fit(rain, 100, show = FALSE)
  ls(fit1)
  fit2 <- oopp.fit(rain, 10, show = FALSE)
  ls(fit2)
}
}
