% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_level.R
\name{plot.retlev}
\alias{plot.retlev}
\title{Plot diagnostics for a ret_lev object}
\usage{
plot.retlev(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{c("ret_lev", "lax")}, a result of
a call to \code{\link{return_level}}.}

\item{y}{Not used.}

\item{...}{Further arguments to be passed to \code{\link[graphics]{plot}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for an objects of class \code{c("ret_lev", "lax")}.
}
\section{Examples}{

See the examples in \code{\link{return_level}}.
}

\seealso{
\code{\link{return_level}}.
}
