% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_level.R
\name{return_level}
\alias{return_level}
\title{Return levels}
\usage{
return_level(x, m = 100, level = 0.95, npy = 1, inc = NULL,
  type = c("vertical", "cholesky", "spectral", "none"))
}
\description{
Add description
}
\examples{
got_evd <- requireNamespace("evd", quietly = TRUE)

if (got_evd) {
  library(evd)
  # An example from the evd::fgev documentation
  uvdata <- evd::rgev(100, loc = 0.13, scale = 1.1, shape = 0.2)
  M1 <- evd::fgev(uvdata)
  adj_fgev <- alogLik(M1)
  summary(adj_fgev)
}

got_ismev <- requireNamespace("ismev", quietly = TRUE)

if (got_ismev) {
  library(ismev)
  # An example from the ismev::gev.fit documentation
  gev_fit <- gev.fit(revdbayes::portpirie, show = FALSE)
  adj_gev_fit <- alogLik(gev_fit)
  return_level(adj_gev_fit)
  ismev::gev.prof(gev_fit, m = 100, xlow = 4.45, xup = 5.5)
}
}
